Project 3 is pretty much done and ProjectB2 is starting up!
Project 3 just needs a file parser and its pretty much done.

ProjectB2 is just beginning and contains nothing as of yet. ProjectB2 is Graham's scan so it would be rather cool if we can show the process on an output. We can use swing to make the form that will display our algorithm along with console output.