The hour is upon us, it looks like this last algorithm won't be finished.
Oh well we can always try and implement while we are in Canada.
This may be the last commit to this repository so lets write a couple more things before we go. The class was fun and I think (hope) I learned a lot. Many of the algorithms were fun to implement and I have a record of all that I did. Hopefully I won't have to manually implement any of these algorithms again, but you never know, some language's librarys can be a bit sparse. That is all for now.